+++ routes/api/users.js
@@ -1,24 +1,17 @@
 const express = require('express');
 const router = express.Router();
 const authController = require('../../controllers/users');
+const auth = require('../../middleware/validateAuth');
 
 router.get('/', authController.listUsers);
 router.post('/signup', authController.signup);
 router.post('/login', authController.login);
-router.get(
-  '/logout',
-  authController.auth,
-  authController.logout
-);
-router.get(
-  '/current',
-  authController.auth,
-  authController.current
-);
+router.get('/logout', auth, authController.logout);
+router.get('/current', auth, authController.current);
 
 router.patch(
   '/:userId/subscription',
-  authController.auth,
+  auth,
   authController.updateSub
 );
